# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: UEMF University Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /courses:
    get:
      operationId: listCourses
      tags: [courses]
      summary: List all courses
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: departmentId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of courses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Course"
    post:
      operationId: createCourse
      tags: [courses]
      summary: Create a new course
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCourseBody"
      responses:
        "201":
          description: Created course
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
  /courses/{id}:
    get:
      operationId: getCourse
      tags: [courses]
      summary: Get a course by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Course
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
    put:
      operationId: updateCourse
      tags: [courses]
      summary: Update a course
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCourseBody"
      responses:
        "200":
          description: Updated course
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
    delete:
      operationId: deleteCourse
      tags: [courses]
      summary: Delete a course
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── FACULTY ──────────────────────────────────────────────────
...
  /departments:
    get:
      operationId: listDepartments
      tags: [departments]
      summary: List all departments
      responses:
        "200":
          description: List of departments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Department"
    post:
      operationId: createDepartment
      tags: [departments]
      summary: Create a department
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDepartmentBody"
      responses:
        "201":
          description: Created department
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"
  /departments/{id}:
    get:
      operationId: getDepartment
      tags: [departments]
      summary: Get a department by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Department
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"
    put:
      operationId: updateDepartment
      tags: [departments]
      summary: Update a department
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDepartmentBody"
      responses:
        "200":
          description: Updated department
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"
    delete:
      operationId: deleteDepartment
      tags: [departments]
      summary: Delete a department
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── ENROLLMENTS ──────────────────────────────────────────────
  /enrollments:
    get:
      operationId: listEnrollments
      tags: [enrollments]
      summary: List all enrollments
      parameters:
        - name: studentId
          in: query
          schema:
            type: integer
        - name: courseId
          in: query
          schema:
            type: integer
        - name: semester
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of enrollments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Enrollment"
    post:
      operationId: createEnrollment
      tags: [enrollments]
      summary: Enroll a student in a course
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEnrollmentBody"
      responses:
        "201":
          description: Created enrollment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Enrollment"
  /enrollments/{id}:
    delete:
      operationId: deleteEnrollment
      tags: [enrollments]
      summary: Remove an enrollment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── GRADES ───────────────────────────────────────────────────
...
  /exams:
    get:
      operationId: listExams
      tags: [exams]
      summary: List all exams
      parameters:
        - name: courseId
          in: query
          schema:
            type: integer
        - name: semester
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of exams
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Exam"
    post:
      operationId: createExam
      tags: [exams]
      summary: Create an exam
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateExamBody"
      responses:
        "201":
          description: Created exam
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exam"
  /exams/{id}:
    put:
      operationId: updateExam
      tags: [exams]
      summary: Update an exam
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateExamBody"
      responses:
        "200":
          description: Updated exam
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exam"
    delete:
      operationId: deleteExam
      tags: [exams]
      summary: Delete an exam
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── FEES ─────────────────────────────────────────────────────
...
  /announcements:
    get:
      operationId: listAnnouncements
      tags: [announcements]
      summary: List all announcements
      responses:
        "200":
          description: List of announcements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Announcement"
    post:
      operationId: createAnnouncement
      tags: [announcements]
      summary: Create an announcement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAnnouncementBody"
      responses:
        "201":
          description: Created announcement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"
  /announcements/{id}:
    put:
      operationId: updateAnnouncement
      tags: [announcements]
      summary: Update an announcement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAnnouncementBody"
      responses:
        "200":
          description: Updated announcement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"
    delete:
      operationId: deleteAnnouncement
      tags: [announcements]
      summary: Delete an announcement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
components:
  schemas:
...
    Course:
      type: object
      properties:
        id:
          type: integer
        courseCode:
          type: string
        name:
          type: string
        description:
          type: string
        credits:
          type: integer
        departmentId:
          type: integer
        departmentName:
          type: string
        facultyId:
          type: integer
        facultyName:
          type: string
        maxStudents:
          type: integer
        enrolledCount:
          type: integer
        semester:
          type: string
        level:
          type: string
          enum: [undergraduate, graduate, doctoral]
        createdAt:
          type: string
      required:
        - id
        - courseCode
        - name
        - credits
        - departmentId
        - semester
        - level
    CreateCourseBody:
      type: object
      properties:
        courseCode:
          type: string
        name:
          type: string
        description:
          type: string
        credits:
          type: integer
        departmentId:
          type: integer
        facultyId:
          type: integer
        maxStudents:
          type: integer
        semester:
          type: string
        level:
          type: string
          enum: [undergraduate, graduate, doctoral]
      required: [courseCode, name, credits, departmentId, semester, level]

    # ─── Faculty ─────────────────────────────────────────────────
...
    Department:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        code:
          type: string
        description:
          type: string
        headFacultyId:
          type: integer
        headFacultyName:
          type: string
        building:
          type: string
        phone:
          type: string
        email:
          type: string
        studentCount:
          type: integer
        facultyCount:
          type: integer
        courseCount:
          type: integer
        createdAt:
          type: string
      required: [id, name, code]
    CreateDepartmentBody:
      type: object
      properties:
        name:
          type: string
        code:
          type: string
        description:
          type: string
        headFacultyId:
          type: integer
        building:
          type: string
        phone:
          type: string
        email:
          type: string
      required: [name, code]

    # ─── Enrollment ──────────────────────────────────────────────
    Enrollment:
      type: object
      properties:
        id:
          type: integer
        studentId:
          type: integer
        studentName:
          type: string
        courseId:
          type: integer
        courseName:
          type: string
        courseCode:
          type: string
        semester:
          type: string
        enrolledAt:
          type: string
        status:
          type: string
          enum: [enrolled, dropped, completed]
      required: [id, studentId, courseId, semester, status]
    CreateEnrollmentBody:
      type: object
      properties:
        studentId:
          type: integer
        courseId:
          type: integer
        semester:
          type: string
      required: [studentId, courseId, semester]

    # ─── Grade ───────────────────────────────────────────────────
...
    Exam:
      type: object
      properties:
        id:
          type: integer
        courseId:
          type: integer
        courseName:
          type: string
        courseCode:
          type: string
        title:
          type: string
        type:
          type: string
          enum: [midterm, final, quiz, practical]
        date:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        room:
          type: string
        semester:
          type: string
        totalMarks:
          type: integer
        instructions:
          type: string
      required:
        - id
        - courseId
        - title
        - type
        - date
        - startTime
        - endTime
        - room
        - semester
        - totalMarks
    CreateExamBody:
      type: object
      properties:
        courseId:
          type: integer
        title:
          type: string
        type:
          type: string
          enum: [midterm, final, quiz, practical]
        date:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        room:
          type: string
        semester:
          type: string
        totalMarks:
          type: integer
        instructions:
          type: string
      required: [courseId, title, type, date, startTime, endTime, room, semester, totalMarks]

    # ─── Fee ─────────────────────────────────────────────────────
...
    Announcement:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        category:
          type: string
          enum: [academic, administrative, event, emergency, general]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        authorId:
          type: integer
        authorName:
          type: string
        targetAudience:
          type: string
          enum: [all, students, faculty, staff]
        publishedAt:
          type: string
        expiresAt:
          type: string
        isActive:
          type: boolean
      required:
        - id
        - title
        - content
        - category
        - priority
        - targetAudience
        - isActive
    CreateAnnouncementBody:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        category:
          type: string
          enum: [academic, administrative, event, emergency, general]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        targetAudience:
          type: string
          enum: [all, students, faculty, staff]
        expiresAt:
          type: string
        isActive:
          type: boolean
      required: [title, content, category, priority, targetAudience]